+++ providers/ic/src/providers/storage/storage.ic.ts
@@ -67,7 +67,7 @@ export const uploadFileIC = async ({
   });
 
   return {
-    downloadUrl: `https://${bucketId.toText()}.raw.ic0.app/${fullPath}?token=${token}`,
+    downloadUrl: `https://${bucketId.toText()}.raw.ic0.app${fullPath}?token=${token}`,
     fullPath,
     name: filename
   };
@@ -89,7 +89,7 @@ export const getFiles: GetFiles = async ({
 
   return {
     items: assets.map(({name, fullPath, token}: AssetKey) => ({
-      downloadUrl: `${host}/${fullPath}?token=${token}`,
+      downloadUrl: `${host}${fullPath}?token=${token}`,
       fullPath,
       name
     })),
@@ -100,8 +100,12 @@ export const getFiles: GetFiles = async ({
 export const deleteFile: DeleteFile = async ({downloadUrl, fullPath}: StorageFile): Promise<void> => {
   const {actor}: BucketActor<StorageBucketActor> = await getStorageActor();
 
-  const {searchParams}: URL = new URL(downloadUrl);
-  const token: string | null = searchParams.get('token');
+  let token: string | null = null;
+
+  if (downloadUrl) {
+    const {searchParams}: URL = new URL(downloadUrl);
+    token = searchParams.get('token');
+  }
 
   return actor.del({fullPath, token: toNullable<string>(token ? token : undefined)});
 };