+++ providers/ic/src/utils/publish.deck.utils.ts
@@ -2,6 +2,7 @@ import {Deck, DeckData, DeckPublishData, deckPublishData, PublishData} from '@de
 
 import {setData} from './data.utils';
 import {initIndexHTML, initUpload, StorageUpload, updateMetaData, uploadPublishFileIC} from './publish.utils';
+import {uploadSocialImage} from './publish.social.utils';
 
 export const publishDeck = async ({
   deck: deckSource
@@ -9,10 +10,11 @@ export const publishDeck = async ({
   deck: Deck;
 }): Promise<{deck: Deck; storageUpload: StorageUpload; publishData: PublishData}> => {
   const {id, data} = deckSource;
+  const {meta} = data;
 
   // 1. Init and fill HTML
   const indexHTML: {html: string; publishData: DeckPublishData} = await initDeckIndexHTML({deck: deckSource});
-  const {storageUpload, publishData} = await initUpload({indexHTML, folder: 'p'});
+  const {storageUpload, publishData} = await initUpload({indexHTML, folder: 'p', meta});
 
   // 2. Update deck published meta
   const deckData: DeckData = updateMetaData<DeckData>({data, meta: data.meta, name: data.name, storageUpload});
@@ -23,7 +25,10 @@ export const publishDeck = async ({
   // 4. Upload
   await uploadPublishFileIC(storageUpload);
 
-  // 5. Tells the snapshot the process is over
+  // 5. Upload
+  await uploadSocialImage({storageUpload, publishData});
+
+  // 6. Tells the snapshot the process is over
   emitDeckPublished(deck);
 
   return {
@@ -34,7 +39,7 @@ export const publishDeck = async ({
 };
 
 const initDeckIndexHTML = async ({deck}: {deck: Deck}): Promise<{html: string; publishData: DeckPublishData}> => {
-  const publishData: DeckPublishData = deckPublishData({deck});
+  const publishData: DeckPublishData = await deckPublishData({deck});
 
   const {slides} = publishData;