+++ providers/ic/src/utils/publish.doc.utils.ts
@@ -2,6 +2,7 @@ import {Doc, DocData, docPublishData, DocPublishData, PublishData} from '@deckde
 
 import {initIndexHTML, initUpload, StorageUpload, updateMetaData, uploadPublishFileIC} from './publish.utils';
 import {setData} from './data.utils';
+import {uploadSocialImage} from './publish.social.utils';
 
 export const publishDoc = async ({
   doc: docSource
@@ -9,10 +10,11 @@ export const publishDoc = async ({
   doc: Doc;
 }): Promise<{doc: Doc; storageUpload: StorageUpload; publishData: PublishData}> => {
   const {id, data} = docSource;
+  const {meta} = data;
 
   // 1. Init and fill HTML
   const indexHTML: {html: string; publishData: DocPublishData} = await initDocIndexHTML({doc: docSource});
-  const {storageUpload, publishData} = await initUpload({indexHTML, folder: 'd'});
+  const {storageUpload, publishData} = await initUpload({indexHTML, folder: 'd', meta});
 
   // 2. Update doc published meta
   const docData: DocData = updateMetaData<DocData>({data, meta: data.meta, name: data.name, storageUpload});
@@ -23,7 +25,10 @@ export const publishDoc = async ({
   // 4. Upload
   await uploadPublishFileIC(storageUpload);
 
-  // 5. Tells the snapshot the process is over
+  // 5. Upload
+  await uploadSocialImage({storageUpload, publishData});
+
+  // 6. Tells the snapshot the process is over
   emitDocPublished(doc);
 
   return {
@@ -34,7 +39,7 @@ export const publishDoc = async ({
 };
 
 const initDocIndexHTML = async ({doc}: {doc: Doc}): Promise<{html: string; publishData: DocPublishData}> => {
-  const publishData: DocPublishData = docPublishData({doc});
+  const publishData: DocPublishData = await docPublishData({doc});
 
   const {paragraphs} = publishData;