+++ providers/ic/src/utils/publish.utils.ts
@@ -4,6 +4,7 @@ import {_SERVICE as StorageBucketActor} from '../canisters/storage/storage.did';
 
 import {BucketActor} from './manager.utils';
 import {encodeFilename, getStorageActor, upload} from './storage.utils';
+import {updateTemplateSocialImage} from './publish.social.utils';
 
 export interface StorageUpload {
   actor: StorageBucketActor;
@@ -26,10 +27,12 @@ export const updateTemplate = ({template, data}: {template: string; data: Partia
 
 export const initUpload = async ({
   indexHTML,
-  folder
+  folder,
+  meta
 }: {
   indexHTML: {html: string; publishData: PublishData};
   folder: 'p' | 'd';
+  meta: Meta | undefined;
 }): Promise<{storageUpload: StorageUpload; publishData: PublishData}> => {
   const {html, publishData} = indexHTML;
 
@@ -37,13 +40,16 @@ export const initUpload = async ({
   const {bucketId, actor}: BucketActor<StorageBucketActor> = await getStorageActor();
 
   // 2. Folder and filename
-  const filename: string = encodeFilename(publishData.title);
-  const pathname: string = `/${folder}/${filename}`;
+  const {filename, pathname} = uploadPaths({publishData, meta, folder});
+
   const bucketUrl: string = `https://${bucketId.toText()}.raw.ic0.app`;
   const fullUrl: string = `${bucketUrl}${pathname}`;
 
   // 3. Update URL
-  const updatedHTML: string = html.replace('{{DECKDECKGO_URL}}', fullUrl);
+  let updatedHTML: string = html.replace('{{DECKDECKGO_URL}}', fullUrl);
+
+  // 4. Update the social image URL
+  updatedHTML = updateTemplateSocialImage({html: updatedHTML, data: publishData, bucketUrl});
 
   return {
     storageUpload: {
@@ -59,6 +65,36 @@ export const initUpload = async ({
   };
 };
 
+/**
+ * !!IMPORTANT!!: The pathname never changes if it has been published once otherwise we cannot delete the content when a doc or deck is deleted
+ */
+const uploadPaths = ({
+  publishData,
+  meta,
+  folder
+}: {
+  publishData: PublishData;
+  folder: 'p' | 'd';
+  meta: Meta | undefined;
+}): {filename: string; pathname: string} => {
+  if (meta?.pathname) {
+    const {pathname} = meta;
+
+    return {
+      filename: pathname.replace(`/${folder}/`, ''),
+      pathname
+    };
+  }
+
+  const filename: string = encodeFilename(publishData.title);
+  const pathname: string = `/${folder}/${filename}`;
+
+  return {
+    filename,
+    pathname
+  };
+};
+
 export const initIndexHTML = async ({
   publishData,
   updateTemplateContent,
@@ -70,7 +106,7 @@ export const initIndexHTML = async ({
 }): Promise<{html: string}> => {
   const template: string = await htmlTemplate(sourceFolder);
 
-  let updatedTemplate: string = updateTemplate({template, data: publishData});
+  const updatedTemplate: string = updateTemplate({template, data: publishData});
 
   const {attributes} = publishData;