+++ studio/src/global/app-ic.ts
@@ -19,6 +19,7 @@ import '@deckdeckgo/word-cloud';
 import '@deckdeckgo/markdown';
 import '@deckdeckgo/laser-pointer';
 import '@deckdeckgo/elements';
+import '@deckdeckgo/social-img';
 
 import '@deckdeckgo/slide-title';
 import '@deckdeckgo/slide-content';
@@ -58,10 +59,10 @@ setupDeckGoConfig({
   },
   cloud: {
     api: {
-      cdn: 'https://unpkg.com/@deckdeckgo/ic@4.0.0/dist/deckdeckgo-ic/index.esm.js'
+      cdn: 'https://unpkg.com/@deckdeckgo/ic@4.1.1/dist/deckdeckgo-ic/index.esm.js'
     },
     signIn: {
-      cdn: 'https://unpkg.com/@deckdeckgo/ic@4.0.0/dist/deckdeckgo-ic/deckdeckgo-ic.esm.js',
+      cdn: 'https://unpkg.com/@deckdeckgo/ic@4.1.1/dist/deckdeckgo-ic/deckdeckgo-ic.esm.js',
       tag: 'deckgo-ic-signin'
     }
   }