+++ providers/ic/CHANGELOG.md
@@ -1,3 +1,17 @@
+# 4.1.1 (2021-12-21)
+
+### Fix
+
+- storage `fullPath` contains begin slash (avoid path like "...ic0.app//images/...")
+
+# 4.1.0 (2021-12-21)
+
+### Features
+
+- publish social images
+- meta pathname fixed once content published
+- handles delete storage without `downloadUrl`
+
 # 4.0.0 (2021-12-17)
 
 ### Breaking Changes